requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.0.6
django-ckeditor==6.3.2
django-debug-toolbar==4.4.2
django-import-export==4.1.0
django-js-asset==2.2.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
pillow==10.3.0
python-dotenv==1.0.1
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.12.2
